+++ tests/columns/test_datecolumn.py
@@ -9,7 +9,7 @@ from django.db import models
 import django_tables2 as tables
 
 
-# Format string: https://docs.djangoproject.com/en/1.4/ref/templates/builtins/#date
+# Format string: https://docs.djangoproject.com/en/stable/ref/templates/builtins/#date
 # D -- Day of the week, textual, 3 letters  -- 'Fri'
 # b -- Month, textual, 3 letters, lowercase -- 'jan'
 # Y -- Year, 4 digits.                      -- '1999'