+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+## unreleased
+- Add tests for TimeColumn
+- 
+
 ## v1.0.7 (2016-01-03)
 - Explicitly check if `column.verbose_name` is not None to support empty column headers (fixes #280)
 - Cleanup the example project to make it work with modern Django versions.